# api/openapi.yaml
openapi: 3.1.0
info:
  title: ForgeERP API
  version: 0.1.0
  description: >
    ForgeERP modernized ERP/CRM API (Dolibarr functional successor).
    This document is authoritative for REST surface; phase routers must register
    every path declared here.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/sepa/batches:
    get:
      operationId: listSepaBatches
      summary: List direct-debit batches (requires sepa.batch.read)
      responses:
        '200': { description: List }
    post:
      operationId: createSepaBatch
      summary: Open a draft batch with IBAN checks (requires sepa.batch.write)
      responses:
        '201': { description: Created }
        '422': { description: Bad IBAN or batch }
...
  /api/v1/sepa/batches/{id}/xml:
    get:
      operationId: sepaXML
      summary: pain.008 export for validated batches (requires sepa.batch.read)
      responses:
        '200': { description: XML }
        '422': { description: Draft batch }
...